Adjust setup to work with CI environment
This changes the expected Docker image to "zeek:latest" and updates the
image-building script to be content with any such existing image, instead of
also checking for fresh local sources. In CI we will have just built a fresh
image anyway.

It also simplifies the way btest invokes "docker exec" to run commands in the
container. The previous approach could unset the PATH set via any ENV in the
Dockerfile.