Support checking "zeek-client get-nodes" for nodes' run-state
This adds a new helper file that provides basic test helpers, the biggie being a
function that will poll "zeek-client get-nodes" to track whether it reports any
cluster nodes, and if so, whether these are all in RUNNING state. We require
this wait-stage to avoid races around cluster startup that can occur when
interacting with cluster nodes that are still booting. This can go away with a
global "cluster is ready" event.

Also set pipefail option to ease use of error detection when piping zeek-client
output into jq.