Avoid discrepancies to standard Zeek Docker image in controller image
In our CI setup the cluster tests run on top of the (freshly built) standard
Zeek docker image, whereas the testsuite uses its own image(s) to assist local
development. We should avoid adding directories that only exist on that image,
because it can make tests fail with one image and succeed with the other.

Also set PYTHONPATH so it automatically finds the Zeek installation's modules.