Fix node state race condition in tests using get-nodes command output
The cluster nodes can sometimes be in state PENDING, not yet RUNNING, which
triggers baseline deviations. This adds a helper script that canonifies the
output at the JSON level. It's much like a btest-diff canonifier, but works best
when running directly on `zeek-client get-nodes` output, possibly followed by
other commands.