docker-compose: Switch to docker compose v2
docker-compose v1 (implemented in Python) has stopped receiving update
in July 2023 and the version installed on Cirrus's docker_builder image
is incompatible with the latest docker-py version, resulting in total
test failures.

This commit switches over the docker-compose invocations to 'docker
compose' as required by docker compose v2. compose is now a plugin of
docker, implemented in Go. It's installable on Ubuntu, for example,
as docker-compose-v2 package.